requirements.txt
einops==0.8.2
matplotlib==3.10.8
numpy==2.4.2
pandas==3.0.0
Pillow==12.1.0
pytorchvideo==0.1.5
scikit_learn==1.8.0
torch==2.6.0+cu124
torch_geometric==2.7.0
torchvision==0.21.0+cu124
tqdm==4.67.1
